% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{create_example}
\alias{create_example}
\title{Function to create a population correlation matrix for the vignette}
\usage{
create_example(p, a, b)
}
\arguments{
\item{p}{total number of variables, must be even}

\item{a}{parameter for the equicorrelation matrix on the diagonal blocks}

\item{b}{parameter for off-diagonal blocks}
}
\value{
a \code{p} x \code{p} matrix.
}
\description{
This function creates a \code{p} x \code{p} matrix where the diagonal blocks are two p/2 equicorrelation matrix with parameter \code{a}. The all the elements in the offdiagonal blocks of the matrix are equal to \code{b}. \code{p} has to be an even number.
}