% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{dCCA}
\alias{dCCA}
\title{Function to evaluate the joint density of the canonical correlations}
\usage{
dCCA(n, p, rp, a, lambda)
}
\arguments{
\item{n}{sample size}

\item{p}{total number of variables}

\item{rp}{smaller of the size of the two groups of variables}

\item{a}{a scaling constant}

\item{lambda}{a \code{rp} length vector where the joint density is to be evaluated \eqn{1 \ge lambda[1]\ge lambda[2] \ge ... \ge lambda[rp] \ge 0}}
}
\value{
Scaled joined density of canonical correlations, evaluated at \eqn{lambda}.
}
\description{
This function evaluates the joint density of the canonical correlations at a specific value \code{lambda}.
}
\keyword{internal}