% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{form_L}
\alias{form_L}
\title{Form L matrix described in Theorem 2}
\usage{
form_L(test_hyp)
}
\arguments{
\item{test_hyp}{output of \code{test_stat_CCA()}}
}
\description{
Form L matrix described in Theorem 2
}
\keyword{internal}