% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{MC_function_classical}
\alias{MC_function_classical}
\title{Function for Monte Carlo simulation for classical inference}
\usage{
MC_function_classical(p, rp, n)
}
\arguments{
\item{p}{\eqn{p_1+p_2}}

\item{rp}{\eqn{min(p_1, p_2)}}

\item{n}{sample size}
}
\value{
A sample from Wilks' lambda distribution.
}
\description{
Samples from the joint distribution of the canonical correlations between two groups of independent Gaussian variables of size \eqn{p_1} and \eqn{p_2} using \code{sample_psi()}, and then computes the corresponding test statistic, which follows a Wilks' lambda distribution.
}
\keyword{internal}