% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{MC_function_selective}
\alias{MC_function_selective}
\title{Function for Monte Carlo simulation for selective inference}
\usage{
MC_function_selective(p, rp, n, L, g)
}
\arguments{
\item{p}{\eqn{p_1 + p_2}}

\item{rp}{\eqn{min(p_1, p_2)}}

\item{n}{sample size}

\item{L}{a matrix with \code{rp} columns}

\item{g}{a vector of length equal to the number of rows in \code{L}}
}
\value{
\item{statistic}{Test statistic corresponding to simulated \eqn{\lambda}.}
\item{status}{A logical vector indicating if the simulated \eqn{\lambda} satisfies \eqn{L\lambda \leq g}.}
}
\description{
Samples from the joint distribution of the canonical correlations between two groups of independent Gaussian variables of size \eqn{p_1} and \eqn{p_2} using \code{sample_psi()}, and then computes the corresponding test statistic, which follows a Wilks' lambda distribution. Given a matrix \code{L} with \code{rp} columns, and a vector of length equal to the number of rows in \code{L}, checks if the sampled vector \eqn{\lambda} satisfies \eqn{\lambda : L\lambda \le g}
}
\keyword{internal}