% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{selective_p_val_beta}
\alias{selective_p_val_beta}
\title{Compute selective p-value using numerical integration}
\usage{
selective_p_val_beta(S, CP, k, n, c, test_hyp)
}
\arguments{
\item{S}{a \eqn{p \times p} covariance matrix}

\item{CP}{a vector of length \eqn{p} with \eqn{i^{th}} element denoting the
group \eqn{i^{th}} variable belongs to}

\item{k}{the group to be tested for independence with the remaining variables, i.e. \eqn{P = [i : CP[i]==k]}}

\item{n}{sample size}

\item{c}{a threshold}

\item{test_hyp}{output of \code{test_stat_CCA()}}
}
\description{
Compute selective p-value using numerical integration
}
\keyword{internal}