% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{test_stat_CCA}
\alias{test_stat_CCA}
\title{Compute the test statistic as a function of canonical correlations.}
\usage{
test_stat_CCA(S, CP, k)
}
\arguments{
\item{S}{a \eqn{p \times p} sample covariance matrix}

\item{CP}{a vector of length \eqn{p} with \eqn{i^{th}} element denoting the
group \eqn{i^{th}} variable belongs to}

\item{k}{the group to be tested for independence with the remaining variables, i.e. \eqn{P = i : CP[i]==k}}
}
\value{
A list containing the following items:
\item{statistic}{Test statistic corresponding to \eqn{S} and group of variables \eqn{P}.}

\item{S11}{\eqn{S_{P, P}} if \eqn{2|P| \ge p}, else \eqn{S_{P^c, P^c}}.}

\item{S22}{\eqn{S_{P^c, P^c}} if \eqn{2|P| \ge p}, else \eqn{S_{P, P}}.}

\item{left_SV}{Left canonical vectors of  \eqn{S_{P, P^c}}.}

\item{right_SV}{Right canonical vectors of \eqn{S_{P, P^c}}.}
}
\description{
Given a sample covariance matrix \eqn{S} and a group of variables \eqn{P}, first
computes the cross-covariance matrix between the whitened variables:
\eqn{S_{P, P^c}^W = S_{P, P}^{-0.5} S_{P, P^c} S_{P^c, P^c}^{-0.5}}. Next,
computes the SVD of \eqn{S_{P, P^c}^W} and returns the test statistic,
\eqn{S_{P, P}}, \eqn{S_{P^c, P^c}} and the canonical vectors
of \eqn{S_{P, P^c}}.
}
\keyword{internal}